#10071: Merge selu_bw to TTNN
