+++ docs/source/ttnn/ttnn/ttnn/selu_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.selu_bw:
+
+  ttnn.selu_bw
+  #############
+
+  .. autofunction:: ttnn.selu_bw
