+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -45,7 +45,8 @@ enum class UnaryBackwardOpType {
     LOG_BW,
     RELU6_BW,
     ABS_BW,
-    SILU_BW
+    SILU_BW,
+    SELU_BW,
 };
 
 struct UnaryBackwardFunction{
